% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_main.R
\name{cptmcmc}
\alias{cptmcmc}
\title{Changepoint for MCMC}
\usage{
cptmcmc(chain, method = c(1, 2), logpost = FALSE, plot = TRUE, thresh = 0.5)
}
\arguments{
\item{chain}{: matrix or vector of the Markov chain. For matrix each column is treated as a component.}

\item{method}{: '1' for local-max method in section 3.1 and 3.2, '2' for modifies as in section 3.3}

\item{logpost}{: F by default, should be T if passing logposterior vector of the chain}

\item{plot}{: T by default}

\item{thresh}{: (0,1). Threshold of maximum viable breakpt fraction.}
}
\value{
Returns a data frame with all the estimatewd breakpts and the truncated chain.
}
\description{
The proposed changepoint algorithms as in the document.
}
\examples{
\dontrun{
cptmcmc(chain_multi,method=1,plot=TRUE)
cptmcmc(chain_log,method=2,logpost=TRUE)
}
}
