% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{MH_multi}
\alias{MH_multi}
\title{MH for multivariate normal}
\usage{
MH_multi(n, start, k, h, mean, sigma)
}
\arguments{
\item{n}{:no. of observations}

\item{start}{: vector of starting values}

\item{k}{: no of dimensions i.e k-variate normal(mean,sigma) is the target}

\item{h}{: vector of step size for MH, i.e proposal will have covariance matrix diag(h)}

\item{mean}{: true target mean}

\item{sigma}{: true target covariance}
}
\value{
Returns a markov chain matrix with dimensions n*k
}
\description{
Generate Markov Chain for target = Nk(mean, sigma), proposal = Nk(x,diag(h))
}
\examples{
\dontrun{
MH_multi(n=1e4, start=c(5,5,5), k=3, h =c(0.1,0.1,0.01), mean=c(0,0,0), sigma=diag(3))
}
}
