% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{MH_multi}
\alias{MH_multi}
\alias{MH_uni}
\title{MH for univariate normal}
\usage{
MH_uni(n, start, h, mean, var)
}
\arguments{
\item{n}{:no. of observations}

\item{start}{: starting value}

\item{h}{: var for proposal}

\item{mean}{: true target mean}

\item{var}{: true target covariance}
}
\value{
Returns a markov chain vector
}
\description{
Generate Markov Chain for target = N(mean, var), proposal = N(x,h)
}
\examples{
\dontrun{
MH_uni(n=1e4, start=5, h=0.1, mean=0, var=1)
}
}
