% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{multinomial_regression}
\alias{multinomial_regression}
\title{Multinomial Regression examples}
\usage{
multinomial_regression(n)
}
\arguments{
\item{n}{: No. of observations required}
}
\value{
Returns a matrix with dimensions n*22.
}
\description{
Generate multinomial regression data
}
\examples{
\dontrun{
multinomial_regression(1e4)}
}
